added validator to messages func in api/__init__ and imported where needed
