+++ app/api/auth_routes.py
@@ -3,20 +3,10 @@ from app.models import User, db
 from app.forms import LoginForm
 from app.forms import SignUpForm
 from flask_login import current_user, login_user, logout_user, login_required
+from . import validation_errors_to_error_messages
 import time
 
 auth_routes = Blueprint("auth", __name__)
-
-
-def validation_errors_to_error_messages(validation_errors):
-    """
-    Simple function that turns the WTForms validation errors into a simple list
-    """
-    errorMessages = []
-    for field in validation_errors:
-        for error in validation_errors[field]:
-            errorMessages.append(f"{field} : {error}")
-    return errorMessages
 
 
 @auth_routes.route("/")
@@ -43,7 +33,7 @@ def login():
         # Add the user to the session, we are logged in!
         user = User.query.filter(User.email == form.data["credential"]).first()
         if not user:
-            user = User.query.filter(User.username == form.data['credential']).first()
+            user = User.query.filter(User.username == form.data["credential"]).first()
         login_user(user)
         return user.to_dict()
     return {"errors": validation_errors_to_error_messages(form.errors)}, 401
