+++ app/api/list_routes.py
@@ -3,6 +3,7 @@ from flask import Blueprint, session, request
 from flask_login import login_required, current_user
 from app.models import List, db
 from app.forms.list_form import ListForm
+from . import validation_errors_to_error_messages
 
 list_routes = Blueprint("lists", __name__)
 
@@ -25,43 +26,47 @@ def create_list():
         db.session.add(li)
         db.session.commit()
         return li.to_dict()
-    return {'errors': validation_errors_to_error_messages(form.errors)}, 401
+    return {"errors": validation_errors_to_error_messages(form.errors)}, 401
+
 
 # UPDATING EXISTING List DETAILS
-@list_routes.route('/', methods=['PATCH'])
+@list_routes.route("/", methods=["PATCH"])
 @login_required
 def update_list():
     userId = current_user.id
-    listId = request.json['listId']
+    listId = request.json["listId"]
 
-    newTitle = request.json['title']
+    newTitle = request.json["title"]
     print(newTitle)
 
     currentList = List.query.get(listId)
     if currentList.userId != userId:
-        return {'list': 'You are not the owner of this list item'}
+        return {"list": "You are not the owner of this list item"}
     currentList.title = newTitle
     currentList.updatedAt = datetime.now()
 
     db.session.commit()
 
-    return {'list': currentList.to_dict()}
+    return {"list": currentList.to_dict()}
+
 
 # # DELETING A List
-@list_routes.route('/', methods=['DELETE'])
+@list_routes.route("/", methods=["DELETE"])
 @login_required
 def del_list():
     # .get() function is essentially 'findByPk()'
     userId = current_user.id
-    listId = request.json['listId']
+    listId = request.json["listId"]
 
     oldList = List.query.get(listId)
     if oldList.userId != userId:
-        return {'list': 'You are not the owner of this list item'}
+        return {"list": "You are not the owner of this list item"}
     db.session.delete(oldList)
     db.session.commit()
 
-    return {'message': 'success'}
+    return {"message": "success"}
+
+
 # =----------------------------------------------------------------------------------------------------
 
 # # IF THERE ARE ANY ISSUES WITH REQUEST BODY, MODIFY
