+++ app/api/task_routes.py
@@ -3,18 +3,7 @@ from flask import Blueprint, request
 from flask_login import login_required, current_user
 from app.models import User, List, Task, Note, db
 from app.forms.task_form import TaskForm
-
-
-def validation_errors_to_error_messages(validation_errors):
-    """
-    Simple function that turns the WTForms validation errors into a simple list
-    """
-    errorMessages = []
-    for field in validation_errors:
-        for error in validation_errors[field]:
-            errorMessages.append(f"{field} : {error}")
-    return errorMessages
-
+from . import validation_errors_to_error_messages
 
 task_routes = Blueprint("task", __name__)
